+++ CHANGELOG.md
@@ -14,7 +14,7 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 
 ## v4.2.1 (WIP)
 
-- Fixed bug not being able to delete projects without first deleting all child
+- Fixed bug where unable to delete a Project without first deleting all child
   objects. (#64)
 
 ## v4.2.0 (2021-09-10)