+++ CHANGELOG.md
@@ -19,7 +19,7 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 
 - Fixed where wharf-core logging for Gin debug and error messages were set up
   after they were initially used, leading to a mix of wharf-core and Gin
-  formatted logs. (#63
+  formatted logs. (#63)
 
 ## v4.2.0 (2021-09-10)