+++ CHANGELOG.md
@@ -21,6 +21,10 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
   after they were initially used, leading to a mix of wharf-core and Gin
   formatted logs. (#63)
 
+- Deprecated endpoint `GET /branch/{branchid}`. Getting a single branch by its
+  ID has been not shown any benefits. Please refer to the
+  `GET /project/{projectid}` endpoint instead. (#75)
+
 ## v4.2.0 (2021-09-10)
 
 - Added support for the TZ environment variable (setting timezones ex.